Update to use `mdast-util-to-hast`
This update ensures less code is in this project: meaning less
chances for bugs to creep in and more updates (as remark-html,
remark-vdom use the same algorithm).

In practice, this changes the output algorithm in two ways:

*   no `<span>`s are inserted around text;
*   newlines are inserted, for readability (mostly harmless, but
    useful when debugging and rendering on the server).

This also adds support for integrations with other remark tools,
through `h()` properties: https://github.com/wooorm/remark-html#integrations

This does however remove support for components to overwrite
`inlineCode`, as this library no longer has knowledge of that.